AgenticAI/ChessApp/requirements.txt
Flask==2.0.1
python-chess==1.999
